# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Open Messenger API
  version: v0.1
  description: HTTP API specification for Open Messenger. The WebSocket endpoint `/v1/events/ws`
    is not included because OpenAPI covers HTTP operations only.
paths:
  /v1/events/stream:
    get:
      summary: Stream Events
      operationId: stream_events_v1_events_stream_get
      responses:
        '200':
          description: Successful Response
          content:
            text/event-stream:
              schema:
                type: string
      tags:
      - Native API
      security:
      - BearerToken: []
...
  /v1/channels:
    post:
      summary: Create Channel
      operationId: create_channel_v1_channels_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
  /v1/channels/{channel_id}:
    get:
      summary: Get Channel
      operationId: get_channel_v1_channels__channel_id__get
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
  /v1/channels/{channel_id}/messages:
    post:
      summary: Create Channel Message
      operationId: create_channel_message_v1_channels__channel_id__messages_post
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NativeCreateMessageRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
    get:
      summary: List Channel Messages
      operationId: list_channel_messages_v1_channels__channel_id__messages_get
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMessagesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
  /v1/messages:batchGet:
    post:
      summary: Batch Get Messages
      operationId: batch_get_messages_v1_messages_batchGet_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchGetMessagesRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchGetMessagesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
  /v1/messages:batchCreate:
    post:
      summary: Batch Create Messages
      operationId: batch_create_messages_v1_messages_batchCreate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCreateMessagesRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchCreateMessagesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
  /v1/channels/{channel_id}/threads:
    post:
      summary: Create Channel Thread
      operationId: create_channel_thread_v1_channels__channel_id__threads_post
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThreadRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
  /v1/threads/{thread_id}/context:
    get:
      summary: Get Thread Context
      operationId: get_thread_context_v1_threads__thread_id__context_get
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: string
          title: Thread Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadContextResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
  /v1/threads/{thread_id}/messages:
    post:
      summary: Create Thread Message
      operationId: create_thread_message_v1_threads__thread_id__messages_post
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: string
          title: Thread Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NativeCreateMessageRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
  /v1/files:
    post:
      summary: Upload File
      operationId: upload_file_v1_files_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_file_v1_files_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileObjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
  /v1/files/{file_id}:
    get:
      summary: Get File
      operationId: get_file_v1_files__file_id__get
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          title: File Id
      responses:
        '200':
          description: Successful Response
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      tags:
      - Native API
      security:
      - BearerToken: []
...
components:
  schemas:
...
    BatchCreateMessagesRequest:
      properties:
        items:
          items:
            $ref: '#/components/schemas/BatchCreateMessageItem'
          type: array
          maxItems: 100
          minItems: 1
          title: Items
      type: object
      required:
      - items
      title: BatchCreateMessagesRequest
...
    BatchGetMessagesRequest:
      properties:
        message_ids:
          items:
            type: string
          type: array
          maxItems: 200
          minItems: 1
          title: Message Ids
      type: object
      required:
      - message_ids
      title: BatchGetMessagesRequest
    BatchGetMessagesResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MessageResponse'
          type: array
          title: Items
        not_found_ids:
          items:
            type: string
          type: array
          title: Not Found Ids
      type: object
      required:
      - items
      - not_found_ids
      title: BatchGetMessagesResponse
...
    ChannelResponse:
      properties:
        channel_id:
          type: string
          title: Channel Id
        name:
          type: string
          title: Name
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - channel_id
      - name
      - created_at
      title: ChannelResponse
    CreateChannelRequest:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
      type: object
      required:
      - name
      title: CreateChannelRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ListMessagesResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MessageResponse'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - items
      - next_cursor
      title: ListMessagesResponse
...
    NativeCreateMessageRequest:
      properties:
        text:
          type: string
          minLength: 1
          title: Text
        thread_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Thread Id
        attachments:
          items:
            type: string
          type: array
          title: Attachments
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        blocks:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Blocks
        mentions:
          items:
            type: string
          type: array
          title: Mentions
        raw_payload:
          additionalProperties: true
          type: object
          title: Raw Payload
      type: object
      required:
      - text
      title: NativeCreateMessageRequest
...
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT-like
      description: 'Use `Authorization: Bearer <token>` for Native API requests. Compatibility
        endpoints that reuse bearer auth may also accept `Authorization: Bot <token>`.'
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Administrative token required by `/admin/v1/*` endpoints.
servers:
- url: http://localhost:8000
  description: Local development
...